Added code to compile some basic Scheme functions.
The compiler now supports display and newline, mostly. See below.
Added a functional test for the compiler, but it is currently disabled. See
below.

I have yet to find a true Gforth equivalent to IBTL's println (or, for now,
Scheme's display and newline). Calling . and cr is close, but each time the top
of the stack is printed with ., a space is appended, making their output not
equivalent.